+++ src/bda/plone/productshop/browser/views.py
@@ -4,6 +4,7 @@ from Acquisition import aq_inner
 from Acquisition import aq_parent
 from bda.plone.ajax.batch import Batch
 from bda.plone.cart.utils import get_object_by_uid
+from bda.plone.orders.interfaces import IBuyable
 from bda.plone.productshop.behaviors import IProductTilesViewSettingsBehavior
 from bda.plone.productshop.interfaces import IProduct
 from bda.plone.productshop.interfaces import IProductGroup
@@ -12,8 +13,9 @@ from bda.plone.productshop.interfaces import IVariant
 from bda.plone.productshop.utils import available_variant_aspects
 from bda.plone.productshop.utils import request_property
 from bda.plone.shop import permissions
-from bda.plone.orders.interfaces import IBuyable
+from plone import api
 from plone.app.contenttypes.behaviors.leadimage import ILeadImage
+from plone.protect.utils import addTokenToUrl
 from plone.registry.interfaces import IRegistry
 from Products.CMFCore.utils import getToolByName
 from Products.Five import BrowserView
@@ -261,6 +263,12 @@ class ProductListingBatch(Batch):
 class ProductListing(BrowserView):
     image_scale = "thumb"
     batch_params = []
+
+    def can_edit(self, context):
+        return api.user.has_permission("Modify portal content", obj=context)
+
+    def edit_url(self, context):
+        return addTokenToUrl(context.absolute_url() + "/edit")
 
     @property
     def slicesize(self):